added overall score filter for diseases targets table
