+++ client/client/app/components/ngbTargetPanel/ngbDiseasesTab/ngbDiseasesTargetsPanel/ngbDiseasesTargetsTableFilter/index.js
@@ -5,7 +5,10 @@ import './ngbDiseasesTargetsTableFilter.scss';
 import component from './ngbDiseasesTargetsTableFilter.component';
 import controller from './ngbDiseasesTargetsTableFilter.controller';
 
-export default angular.module('ngbDiseasesTargetsTableFilter', [])
+import filterInput from './ngbTargetsFilterInput';
+import filterList from './ngbTargetsFilterList';
+
+export default angular.module('ngbDiseasesTargetsTableFilter', [filterList, filterInput])
     .controller(controller.UID, controller)
     .component('ngbDiseasesTargetsTableFilter', component)
     .name;
