+++ client/client/app/components/ngbTargetPanel/ngbDiseasesTab/ngbDiseasesTargetsPanel/ngbDiseasesTargetsTableFilter/ngbTargetsFilterList/ngbTargetsFilterList.component.js
@@ -1,0 +1,9 @@
+import controller from './ngbTargetsFilterList.controller';
+
+export default  {
+    bindings: {
+        column: '<'
+    },
+    controller: controller.UID,
+    template: require('./ngbTargetsFilterList.tpl.html')
+};
